imdb_sentiment_analysis/requirements.txt
torch
transformers
dataset
accelerate
evaluate
scikit-learn
humpy
sympy==1.12